fix 32-bit signed int at one spot in the Miller JSON reader

+++ c/input/json_parser.h
@@ -146,7 +146,7 @@ json_value_t * json_parse(
 	size_t length,
 	char * error_buf,
 	json_char** ppend_of_item,
-	int *pline_number); // should be set to 0 by the caller before 1st call
+	long long *pline_number); // should be set to 0 by the caller before 1st call
 
 json_value_t * json_parse_for_unit_test(
 	const json_char * json,
